+++ solver_rust_conic/totsu_core/CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `LinAlgEx::add_sub` for performance improvement.
+- `LinAlg::max` for performance improvement.
 ### Changed
 - Performance improvement of `ConeRotSOC::proj`.
+- Performance improvement of `ConeRPos::proj`.
 - Performance improvement of `SliceLike::get` and `SliceLike::set`.
 ### Deprecated
 ### Removed