+++ solver_rust_conic/totsu_f32cuda/CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `F32CUDA::add_sub` for performance improvement.
+- `F32CUDA::max` for performance improvement.
+- `kernel.ptx` and `kernel.cu`.
 ### Changed
+- `CudaHandle` including context, stream and modlue.
 ### Deprecated
 ### Removed
 ### Fixed